+++ packages/components/list-auto-grid.d.ts
@@ -1,0 +1,1 @@
+export * from './dist/list-auto-grid';