+++ packages/components/src/list-auto-grid/index.ts
@@ -1,0 +1,5 @@
+import ListAutoGrid from './src/index.vue';
+
+export {
+  ListAutoGrid as MixteListAutoGrid,
+};